clean up

+++ letterboxd.sql
@@ -102,7 +102,7 @@ DROP TABLE IF EXISTS `unwosu`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `unwosu` (
   `Name` text DEFAULT NULL,
-  `Year` bigint(20) DEFAULT NULL,
+  `Year` double DEFAULT NULL,
   `Rating` double DEFAULT NULL,
   `Meter_Score` double DEFAULT NULL,
   `Difference` double DEFAULT NULL,
@@ -129,4 +129,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2021-07-09 17:26:01
+-- Dump completed on 2021-07-09 18:22:23
